cleaning up various files
