+++ huntserver/staff_views.py
@@ -7,7 +7,7 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 from .models import Submission, Hunt, Team, Puzzle, Unlock, Solve, Message
 from .forms import SubmissionForm, UnlockForm
-from .puzzle import unlock_puzzles, download_puzzles
+from .utils import unlock_puzzles, download_puzzles
 
 @staff_member_required
 def queue(request, page_num=1):
@@ -26,13 +26,13 @@ def queue(request, page_num=1):
     else:
         hunt = Hunt.objects.get(hunt_number=settings.CURRENT_HUNT_NUM)
         submissions = Submission.objects.filter(puzzle__hunt=hunt).select_related('team', 'puzzle').order_by('-pk')
-        paginator = Paginator(submissions, 30)
+        pages = Paginator(submissions, 30)
         try:
-            submissions = paginator.page(page_num)
+            submissions = pages.page(page_num)
         except PageNotAnInteger:
-            submissions = paginator.page(1)
+            submissions = pages.page(1)
         except EmptyPage:
-            submissions = paginator.page(paginator.num_pages)
+            submissions = pages.page(pages.num_pages)
         form = SubmissionForm()
         context = {'form': form, 'submission_list': submissions}
         return render(request, 'queue.html', context)
